+++ kitty/tabs.py
@@ -572,7 +572,8 @@ class Tab:  # {{{
         if group is not None:
             w = self.active_window
             if w is not None and w.id != window_id:
-                self.current_layout.move_window_to_group(self.windows, group.id)
+                if self.current_layout.move_window_to_group(self.windows, group.id):
+                    self.relayout()
 
     @ac('win', '''
         Focus a visible window by pressing the number of the window. Window numbers are displayed
@@ -583,6 +584,12 @@ class Tab:  # {{{
             tab.set_active_window(window)
 
         get_boss().visual_window_select_action(self, callback, 'Choose window to switch to')
+
+    @ac('win', 'Swap the current window with another window in the current tab, selected visually')
+    def swap_with_window(self) -> None:
+        def callback(tab: Tab, window: Window) -> None:
+            tab.swap_active_window_with(window.id)
+        get_boss().visual_window_select_action(self, callback, 'Choose window to swap with')
 
     @ac('win', 'Move active window to the top (make it the first window)')
     def move_window_to_top(self) -> None: