+++ docs/changelog.rst
@@ -32,7 +32,7 @@ To update |kitty|, :doc:`follow the instructions <binary>`.
   the request by default instead of denying by default. See
   :opt:`clipboard_control` for details (:iss:`4022`)
 
-- A new mappable action ``select_window_in_tab`` to select a window in the current tab to switch to, by window title
+- A new mappable action ``swap_with_window`` to swap the current window with another window in the tab, visually
 
 - A new option :opt:`bell_path` to specify the path to a sound file
   to use as the bell sound